Finalize merge by using Sidebar instead of Navbar component
Files are also organized by "Views" and tsx files have their own css files
