+++ client/src/components/Sidebar/Sidebar.tsx
@@ -1,12 +1,14 @@
+import { Link } from "react-router-dom";
 import { SidebarProps } from "../../types/types";
 import { SidebarOption } from "./SidebarOption";
 import "./Sidebar.css";
 
 const optionStrings = [
-  "Start Workout",
-  "My Workouts",
-  "Track Progress",
-  "Change Preferences",
+  { str: "Start Workout", path: "/" },
+  { str: "My Workouts", path: "/workout-planner" },
+  { str: "Exercise Library", path: "/exercise-library" },
+  { str: "Track Progress", path: "/" },
+  { str: "Change Preferences", path: "/" },
 ];
 
 export const Sidebar = (props: SidebarProps) => {
@@ -14,18 +16,19 @@ export const Sidebar = (props: SidebarProps) => {
     <div className="sidebar">
       <div className="sidebar-container pt-3">
         <p className="sidebar-title fs-2 mb-3">Triton Workout Planner</p>
-        {optionStrings.map((option: string, id: number) => {
+        {optionStrings.map((option, id: number) => {
           return (
-            <div
+            <Link
+              to={option.path}
               key={id}
               className="sidebar-option"
               onClick={() => props.sidebarClickHandler(id)}
             >
               <SidebarOption
-                optionString={option}
+                optionString={option.str}
                 isCurrentPage={id === props.pageIndex}
               />
-            </div>
+            </Link>
           );
         })}
       </div>
