+++ client/src/components/Sidebar/SidebarOption.css
@@ -12,6 +12,7 @@
   padding-bottom: 0.5rem;
   margin-bottom: 0;
   border-radius: 20px;
+  color: black;
 }
 
 .sidebarOption-container p:hover:not(.sidebarOption-current) {
