--- client/src/components/WorkoutPlanner.tsx
@@ -1,19 +1,0 @@
-import '../App.css';
-import { Link } from "react-router-dom";
-import React from 'react';
-
-export const WorkoutPlanner = () => {
-    return (
-        <div className="workout-planner-main">
-            <h1>Choose a Plan</h1>
-            <div className="workout-planner-options">
-                <Link to="/build-your-own" className="planner-link">
-                    Customize Your Own
-                </Link>
-                <Link to="/generate-workout" className="planner-link">
-                    Select Existing Workout Plan
-                </Link>
-            </div>
-        </div>
-    );
-}
