+++ client/src/views/BuildYourOwn.css
@@ -1,0 +1,57 @@
+.build-your-own {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  height: full;
+}
+
+.build-your-own iframe {
+  width: 650px;
+  height: 500px;
+  margin-left: 20px;
+  border: 2px solid #4c566a;
+  border-radius: 8px;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+}
+
+select {
+  width: 200px;
+  padding: 10px;
+  margin: 8px 0;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  font-size: 16px;
+  background-color: #f8f8f8;
+  transition: border-color 0.3s ease;
+}
+
+select:focus {
+  border-color: #4caf50;
+  outline: none;
+}
+
+.day-selection {
+  margin-bottom: 16px;
+  white-space: nowrap;
+}
+
+.day-selection label {
+  margin-right: 10px;
+  white-space: nowrap;
+}
+
+.build-your-own button {
+  background-color: #4caf50;
+  color: white;
+  padding: 10px 15px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 16px;
+  transition: background-color 0.3s ease;
+}
+
+.generated-plan {
+  align-items: center;
+  display: center;
+}
