+++ client/src/views/ExerciseLibrary.css
@@ -1,0 +1,34 @@
+.exercise-library {
+  background-color: #eceff4;
+}
+
+.exercise-library h1 {
+  color: #2e3440;
+}
+
+.exercise-item h3 {
+  color: #d8dee9;
+}
+
+.exercise-item {
+  background-color: #4c566a;
+  margin: 20px;
+  padding: 10px;
+  text-align: center;
+  width: 150px;
+  height: 60px;
+  border: 1px solid #ccc;
+  border-radius: 10px;
+  align-items: center;
+  justify-content: center;
+  word-wrap: break-word;
+  display: flex;
+}
+
+.exercises-list {
+  margin-left: 100px;
+  margin-right: 100px;
+  align-items: center;
+  justify-content: center;
+  display: flex;
+}
