--- client/src/views/Homepage.tsx
@@ -1,14 +1,0 @@
-import { useState } from "react";
-import { Sidebar } from "../components/Sidebar/Sidebar";
-
-export const Homepage = () => {
-  const [pageIndex, setPageIndex] = useState(0);
-
-  function handleSidebarClick(pageIndex: number) {
-    setPageIndex(pageIndex);
-  }
-
-  return (
-    <Sidebar pageIndex={pageIndex} sidebarClickHandler={handleSidebarClick} />
-  );
-};
